Fix restore for devices that don't have eUICC
Because of a default value of (uint64_t)-1LL returned when _plist_dict_get_uint
doesn't find the dictionary entry for the given key, a later comparison of
that unsigned value against something like >= 5 will result in the condition
being true even though it was not supposed to. _plist_dict_get_uint will now
return a default value of 0 if the key is not found. Code paths that deal with
actual values of 0 vs. non-existent values need to test the existence of the
key deal with that; I am currently not aware of anything that would be affected.